% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mRNAcal.R
\name{run_mRNAcal}
\alias{run_mRNAcal}
\title{Run the mRNA Analysis Pipeline}
\usage{
run_mRNAcal(
  genbank_file = NULL,
  rnafold_path = "/Users/JaeYoon/miniconda3/bin/RNAfold",
  upstream_values = seq(100, 500, 100),
  downstream = 100,
  custom_downstream_seq = NULL
)
}
\arguments{
\item{genbank_file}{A character string specifying the path to the GenBank file. Default is \code{NULL}.
If \code{NULL}, the function uses the preloaded \code{genbank_table} in the environment.}

\item{rnafold_path}{A character string specifying the path to the RNAfold executable. Default is
\code{"/Users/JaeYoon/miniconda3/bin/RNAfold"}.}

\item{upstream_values}{A numeric vector specifying the upstream lengths to extract for mRNA regions.
Default is \code{seq(100, 500, 100)}.}

\item{downstream}{A numeric value specifying the downstream length to extract for mRNA regions.
Default is \code{100}.}

\item{custom_downstream_seq}{A character string specifying a custom downstream sequence for mRNA regions.
If provided, it overrides the downstream parameter. Default is \code{NULL}.}
}
\value{
The processed data frame with RNAfold results appended.
}
\description{
This function executes a series of steps in the GenBank analysis pipeline, including
organizing GenBank data, extracting sequences, and processing RNAfold results.
Progress messages are displayed at each step to provide updates.
}
\details{
This function performs the following steps:
\enumerate{
\item Organizes GenBank data using \code{Genbank_organizer}.
\item Extracts FASTA sequences with \code{genbank_fna_extractor}.
\item Retrieves basic GenBank information using \code{gb_info}.
\item Extracts mRNA regions with \code{generate_mRNA_regions}.
\item Processes RNAfold results using \code{process_all_sequences}.
}
Progress messages are displayed at each step to inform the user of the pipeline status.
}
\examples{
# Example usage
run_mRNAcal(
  genbank_file = "example.gbk",
  rnafold_path = "/path/to/RNAfold",
  upstream_values = seq(100, 500, 100),
  downstream = 200,
  custom_downstream_seq = NULL
)

}
