% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mRNAfold.R
\name{run_mRNAfold}
\alias{run_mRNAfold}
\title{Execute mRNAfold on a Given RNA Sequence}
\usage{
run_mRNAfold(sequence, rnafold_path = "/Users/JaeYoon/miniconda3/bin/RNAfold")
}
\arguments{
\item{sequence}{A character string representing the RNA sequence to fold.}

\item{rnafold_path}{A character string specifying the path to the RNAfold executable.
Default is "/usr/local/bin/RNAfold".}
}
\value{
A list containing the following components:
\describe{
\item{mRNA_Sequence}{The RNA sequence as processed by RNAfold.}
\item{Structure}{The dot-bracket structure of the RNA.}
\item{FreeEnergy}{The minimum free energy (ΔG) value of the RNA structure.}
}
}
\description{
This function runs RNAfold on a given RNA sequence and extracts the dot-bracket structure
and minimum free energy (ΔG) from the RNAfold output.
}
\details{
The function assumes that RNAfold is installed and available at the specified path.
It captures the RNAfold output and parses the dot-bracket structure and ΔG value.
If RNAfold output is malformed or incomplete, the function throws an error.
}
\examples{
# Example usage:
sequence <- "GCGCUUCGCCGAAGCGCUUCGCCGAAGCGCUUCGCCGAAGCGCUUCGCCGAAGC"
result <- run_rnafold(sequence)
print(result)

}
